% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linpk.R
\name{as.data.frame.pkprofile}
\alias{as.data.frame.pkprofile}
\title{Coerse a \code{pkprofile} to a \code{data.frame}}
\usage{
\method{as.data.frame}{pkprofile}(x, ...)
}
\arguments{
\item{x}{An object of class \code{pkprofile}.}

\item{...}{Further arguments passed along.}
}
\value{
A \code{data.frame} with columns \code{time} and \code{conc}.
}
\description{
Coerse a \code{pkprofile} to a \code{data.frame}
}

