% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linpk.R
\name{cor2cov}
\alias{cor2cov}
\title{Convert from standard deviation and correlation matrix to covariance matrix.}
\usage{
cor2cov(cor, sd)
}
\arguments{
\item{cor}{A correlation matrix. If \code{sd} is missing, the diagonal
entries are taken to be the standard deviations, otherwise they are ignored.}

\item{sd}{A vector of standard deviations (optional).}
}
\value{
A covariance matrix.
}
\description{
Convert from standard deviation and correlation matrix to covariance matrix.
}
\examples{
cor2cov(matrix(c(1, 0.5, 0.5, 1), 2, 2), 0.1)
}
