% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linpk.R
\name{Diag}
\alias{Diag}
\title{Diagonal Matrix}
\usage{
Diag(x, ..., .names = names(x), .colnames = .names, .rownames = .names)
}
\arguments{
\item{x}{A numeric (or number-like) vector (possibly named).}

\item{...}{Additional numeric (or number-like) vectors (possibly named).}

\item{.names, .colnames, .rownames}{Optionally, specify the row and column
names of the resulting diagonal matrix.}
}
\description{
Like the base \code{\link{diag}} function, except that vectors of length
one are converted to 1-by-1 matrices, values can be specified either as a
single vector argument or multiple arguments, and row and column names can
be specified.
}
\examples{
Diag(6)
Diag(3.14, .names="pi")
Diag(1:6, .colnames=LETTERS[1:6], .rownames=letters[1:6])
Diag(1, 2, 3)
Diag(a=1, b=2, c=3)
Diag(a=1, c(b=2, c=3))
Diag(2+3i, 4+5i)
Diag(T, F)
}
