% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linpk.R
\name{finalstate}
\alias{finalstate}
\title{Get the final state of a linear PK system.}
\usage{
finalstate(x)
}
\arguments{
\item{x}{A object of class \code{\link{pkprofile}}.}
}
\value{
A \code{numeric} vector containing the state of each compartment at
the final observation time.
}
\description{
Get the final state of a linear PK system.
}
\examples{
# Administer a dose at time 0 and a second dose using the final state
# from the first dose (at 12h) as the initial state for the second dose.
t.obs <- seq(0, 12, 0.1)
y <- pkprofile(t.obs, cl=0.25, vc=5, ka=1, dose=list(t.dose=0, amt=1))
finalstate(y)
y2 <- pkprofile(t.obs, cl=0.25, vc=5, ka=1, dose=list(t.dose=0, amt=1), initstate=finalstate(y))
plot(y, xlim=c(0, 24), ylim=c(0, max(y2)), col="blue")  # First dose
lines(t.obs+12, y2, col="red")                          # Second dose
}

