% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linpk.R
\name{linpkApp}
\alias{linpkApp}
\title{Runs the interactive shiny app.}
\usage{
linpkApp(...)
}
\arguments{
\item{...}{Arguments passed to \code{shiny::runApp()}.}
}
\value{
Called for its side effects.
}
\description{
Runs the interactive shiny app.
}
\examples{
\dontrun{
linpkApp()
}
}
