% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linpk.R
\name{secondary}
\alias{secondary}
\title{Derive secondary PK parameters.}
\usage{
secondary(x)
}
\arguments{
\item{x}{A object of class \code{\link{pkprofile}}.}
}
\value{
A \code{list} containing the following:
\describe{
  \item{\code{Cmax}}{Maximum concentration over the interval between doses.}
  \item{\code{Tmax}}{Time of the maximum concentration over the interval between doses.}
  \item{\code{Cmin}}{Minimum concentration over the interval between doses.}
  \item{\code{Tmin}}{Time of the minimum concentration over the interval between doses.}
  \item{\code{Ctrough}}{Concentration at the time of each dose (i.e. just prior to the dose).}
  \item{\code{Ttrough}}{Time of each dose.}
  \item{\code{Cave}}{Average concentration over the interval between doses
  (calculated by the trapezoid rule).}
  \item{\code{AUC}}{Area under the concentration-time curve over the
  interval between doses (calculated by the trapezoid rule).}}
}
\description{
Derive secondary PK parameters.
}
\examples{
t.obs <- seq(0, 24*4, 0.1)
y <- pkprofile(t.obs, cl=0.25, vc=5, ka=1, dose=list(t.dose=0, amt=1, addl=6, ii=12))
secondary(y)
plot(y)
with(secondary(y), points(Tmax, Cmax, pch=19, col="blue"))
with(secondary(y), points(Tmin, Cmin, pch=19, col="red"))
with(secondary(y), points(Ttrough, Ctrough, pch=19, col="green"))
with(secondary(y), points(Ttrough + 6, Cave, pch=19, col="purple", cex=2))

}

