% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linpk.R
\name{blockdiag}
\alias{blockdiag}
\title{Construct a block-diagonal matrix.}
\usage{
blockdiag(...)
}
\arguments{
\item{...}{Any number of square matrices making up the diagonal blocks of
the matrix.}
}
\value{
A block-diagonal matrix.
}
\description{
Construct a block-diagonal matrix.
}
\examples{
blockdiag(matrix(1, 2, 2), 2, matrix(3, 4, 4))
}
