% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linpk.R
\name{linpkApp}
\alias{linpkApp}
\title{Runs the interactive shiny app.}
\usage{
linpkApp(...)
}
\arguments{
\item{...}{Arguments passed to \code{shiny::runApp()}.}
}
\value{
Called for its side effects.
}
\description{
Runs the interactive shiny app.
}
\section{Note}{

The app requires the following packages:
\itemize{
  \item `shiny`
  \item `shinyjs`
  \item `shinyAce`
  \item `dygraphs`
}
Make they are installed or the app won't work.
}

\examples{
\dontrun{
linpkApp()
}
}
