% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linpk.R
\name{LTmat}
\alias{LTmat}
\title{Construct a symmetric matrix from its lower triangle.}
\usage{
LTmat(LT)
}
\arguments{
\item{LT}{A numeric vector giving the elements of the lower triangle of the
matrix by row (see examples).}
}
\value{
A symmetric matrix.
}
\description{
Construct a symmetric matrix from its lower triangle.
}
\examples{
LTmat(1:6)
}

