% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linpk.R
\name{pkprofile-methods}
\alias{pkprofile-methods}
\alias{print.pkprofile}
\alias{plot.pkprofile}
\alias{lines.pkprofile}
\alias{points.pkprofile}
\title{Printing and plotting methods for class \code{pkprofile}.}
\usage{
\method{print}{pkprofile}(x, ...)

\method{plot}{pkprofile}(x, y, ...)

\method{lines}{pkprofile}(x, y, ...)

\method{points}{pkprofile}(x, y, ...)
}
\arguments{
\item{x}{An object of class \code{pkprofile}.}

\item{y}{Any other object. Specifying \code{y} causes the default method to
be called instead (effectively overriding the class-specific behaviour).}
}
\description{
Printing and plotting methods for class \code{pkprofile}.
}
\keyword{internal}
